+++ app/controllers/api/v1/subjects_controller.rb
@@ -3,19 +3,23 @@ class Api::V1::SubjectsController < ApplicationController
 
   # Authenticated
   def create
-    game = Game.find_by_id(game_params[:game_id])
-    if api_v1_user_signed_in? && current_api_v1_user.id == game.user_id
-      invalid_words = []
-      words_params.each do |word|
-        subject = Subject.create_or_find_by(game_id: game_params[:game_id], word: word.upcase)
-        invalid_words << word unless subject.save
-      end
-      if invalid_words.length == 0
-        render json: { isLoggedIn: true, ok: true, message: "All of the words have been created successfully." }, status: 201
-      elsif words_params.length > invalid_words.length
-        render json: { isLoggedIn: true, ok: true, message: "The words have been created successfully, but some of them were failed.", data: invalid_words }, status: 207
+    if api_v1_user_signed_in?
+      game = Game.find_by_id(game_params[:game_id])
+      if game.present? && current_api_v1_user == game.owner
+        invalid_words = []
+        words_params.each do |word|
+          subject = Subject.create_or_find_by(game_id: game_params[:game_id], word: word.upcase)
+          invalid_words << word unless subject.save
+        end
+        if invalid_words.length == 0
+          render json: { isLoggedIn: true, ok: true, message: "All of the words have been created successfully." }, status: 201
+        elsif words_params.length > invalid_words.length
+          render json: { isLoggedIn: true, ok: true, message: "The words have been created successfully, but some of them were failed.", data: invalid_words }, status: 207
+        else
+          render json: { isLoggedIn: true, ok: false, message: "The words are failed." }, status: 500
+        end
       else
-        render json: { isLoggedIn: true, ok: false, message: "The words are failed." }, status: 500
+        render json: { isLoggedIn: false, ok: false, message: "You are not the owner of the game." }, status: 401
       end
     else
       render json: { isLoggedIn: false, ok: false, message: "You are not logged in." }, status: 401
@@ -25,10 +29,14 @@ class Api::V1::SubjectsController < ApplicationController
   def update
     if api_v1_user_signed_in?
       subject = Subject.find_by_id(subject_params[:id])
-      if subject.update(subject_params)
-        render json: { isLoggedIn: true, ok: true, message: "Updated.", data: subject }, status: 200
+      if subject.present? && subject.game.owner == current_api_v1_user
+        if subject.update(subject_params.to_h.transform_values{|v| v.is_a?(String) ? v.upcase : v})
+          render json: { isLoggedIn: true, ok: true, message: "Updated.", data: subject }, status: 200
+        else
+          render json: { isLoggedIn: true, ok: false, message: subject.errors.messages, data: subject }, status: 500
+        end
       else
-        render json: { isLoggedIn: true, ok: false, message: "The parameter is incorrect." }, status: 500
+        render json: { isLoggedIn: false, ok: false, message: "You are not the owner of the game." }, status: 401
       end
     else
       render json: { isLoggedIn: false, ok: false, message: "You are not logged in." }, status: 401
@@ -38,10 +46,14 @@ class Api::V1::SubjectsController < ApplicationController
   def destroy
     if api_v1_user_signed_in?
       subject = Subject.find_by_id(params[:id])
-      if subject.destroy
-        render json: { isLoggedIn: true, ok: true, message: "Deleted.", data: subject }, status: 200
+      if subject.present? && subject.game.owner == current_api_v1_user
+        if subject.destroy
+          render json: { isLoggedIn: true, ok: true, message: "Deleted.", data: subject }, status: 200
+        else
+          render json: { isLoggedIn: true, ok: false, message: "Failed." }, status: 500
+        end
       else
-        render json: { isLoggedIn: true, ok: false, message: "Failed." }, status: 500
+        render json: { isLoggedIn: false, ok: false, message: "You are not the owner of the game." }, status: 401
       end
     else
       render json: { isLoggedIn: false, ok: false, message: "You are not logged in." }, status: 401
